+++ lib/glimmer/libui/arc.rb
@@ -28,7 +28,7 @@ module Glimmer
       parameter_defaults 0, 0, 0, 0, 360, false
               
       def draw(area_draw_params)
-        @args[5] ||= ControlProxy.boolean_to_integer(@args[5], allow_nil: false)
+        @args[5] ||= Glimmer::LibUI.boolean_to_integer(@args[5], allow_nil: false)
         if parent.is_a?(Figure) && parent.x.nil? && parent.y.nil?
           ::LibUI.draw_path_new_figure_with_arc(path_proxy.libui, *@args)
         else