+++ lib/glimmer/libui/grid_proxy.rb
@@ -35,7 +35,7 @@ module Glimmer
         child.halign = 0 if child.halign.nil?
         child.vexpand = false if child.vexpand.nil?
         child.valign = 0 if child.valign.nil?
-        ::LibUI.grid_append(@libui, child.libui, child.left, child.top, child.xspan, child.yspan, ControlProxy.boolean_to_integer(child.hexpand), child.halign, ControlProxy.boolean_to_integer(child.vexpand), child.valign)
+        ::LibUI.grid_append(@libui, child.libui, child.left, child.top, child.xspan, child.yspan, Glimmer::LibUI.boolean_to_integer(child.hexpand), child.halign, Glimmer::LibUI.boolean_to_integer(child.vexpand), child.valign)
         children << child
       end