+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support `area` `transform` property for building `matrix` objects (either through `area { transform {operations} }` directly or through `m1 = matrix {operations}; area { transform m1 }` to reuse a matrix)
 - Ensure `matrix` `rotate` method accepts degrees (not radians)
 - Support `matrix` `multiply` method that accepts a `Glimmer::LibUI::MatrixProxy` object
+- Support `matrix` `invertible?` property that returns a Boolean
 
 ## 0.1.4